Reorganize some utilities
