+++ cli/scripts/execute.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../src/cli').run();
+require('../src').run();
