+++ cli/src/util/__tests__/printErrors.test.js
