+++ cli/src/util/printErrors.js
